+++ change/@fluentui-react-overflow-7caf7721-7aaf-4793-9640-e9844c06d0d8.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Expose Overflow use/render/styles hooks, provider and context types for headless consumers",
+  "packageName": "@fluentui/react-overflow",
+  "email": "vgenaev@gmail.com",
+  "dependentChangeType": "patch"
+}
