+++ packages/react-components/react-overflow/library/src/index.ts
@@ -1,19 +1,35 @@
-export { Overflow } from './components/Overflow';
-export type { OverflowProps, OnOverflowChangeData } from './components/Overflow';
+export {
+  Overflow,
+  useOverflow_unstable,
+  renderOverflow_unstable,
+  useOverflowStyles_unstable,
+} from './components/Overflow';
+export type {
+  OverflowProps,
+  OnOverflowChangeData,
+  OverflowState,
+  OverflowComponentState,
+  OverflowContextValues,
+} from './components/Overflow';
 export { DATA_OVERFLOWING, DATA_OVERFLOW_ITEM, DATA_OVERFLOW_MENU, DATA_OVERFLOW_DIVIDER } from './constants';
 export type { UseOverflowContainerReturn } from './types';
 export { useIsOverflowGroupVisible } from './useIsOverflowGroupVisible';
 export { useIsOverflowItemVisible } from './useIsOverflowItemVisible';
 export { useOverflowContainer } from './useOverflowContainer';
+export { useOverflowContextValues_unstable } from './useOverflowContextValues';
 export { useOverflowCount } from './useOverflowCount';
 export { useOverflowItem } from './useOverflowItem';
 export { useOverflowMenu } from './useOverflowMenu';
 export { useOverflowDivider } from './useOverflowDivider';
 export { useOverflowVisibility } from './useOverflowVisibility';
 
-export { useOverflowContext } from './overflowContext';
+export { useOverflowContext, OverflowProvider } from './overflowContext';
+export type { OverflowContextValue } from './overflowContext';
 
 export type { OverflowItemProps } from './components/OverflowItem/OverflowItem.types';
+export type { OverflowDividerProps } from './components/OverflowDivider/OverflowDivider.types';
 export { OverflowItem } from './components/OverflowItem/OverflowItem';
 export { OverflowDivider } from './components/OverflowDivider/OverflowDivider';
 export { OverflowReorderObserver } from './components/OverflowReorderObserver';
+
+export type { OnUpdateItemVisibility, OnUpdateOverflow, OverflowEventPayload } from '@fluentui/priority-overflow';
